+++ examples/starter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @liteforge/starter
 
+## 0.0.29
+
+### Patch Changes
+
+- liteforge@0.7.9
+
 ## 0.0.28
 
 ### Patch Changes